requirements.txt
# data_science_utils.egg==info
#cython 
#gdal
geopandas==0.10
pandas==1.5.3
pydeck==0.8.1b0
python-dotenv==1.0.0
Shapely==2.0
streamlit==1.24
varname==0.11.2
